Display artifacts that are in the folder
This is part of story #9468 import folders

As soon as a link '_is_folder' is done between some artifacts and a
folder, when you go on the "Folder" tab of the latter, then you can see
the artifacts. This is the same table as what is displayed in the
artifact link nature tab + a column "Folder". This column will be filled
later.

If you don't have permissions to read an artifact, it is not displayed.

Change-Id: I311a1aacb3d89537f1e61ca6f7b846608d771102