REST route for kanban cumulative flow
You now have a REST route returning the state of the kanban columns for a given period of
time.

This is part of story #9463 see cumulative flow diagram

The way the data for the REST tests has been a little bit modified so we can load data
of the plugins after the core.

Change-Id: Id8ae9d583ddbfa60cdfda46aa03f21fd7680f4ba