Display hierarchy of folders in column
part of story #9468 import folders

Fill the Folder column with the parents of current folders.
If the hierarchy is a graph (more than one parent at a given level) then
only one parent is taken (the warning is already displayed at the top of
the page).

Only "Folders" from the hierarchy are displayed.
When you click on the folder you go directly to the Folder pane.

The column is filled only for top level artifacts. We will fill the
column for children in a dedicated changeset.

Change-Id: Ifa6f1beec695fcc33b7279b7e97fc5127a6e7d1a