Fix script for release generation
When we change the version number, we must first commit with VERSION updated
but the script won't properly generate the package number because the tag is
not available yet.

Then if we detect that the last tag is different from the VERSION file, we
decide it's a "-1" release.

+ Minor fix, as when there is no commit after the version, the package is -1,
then when there is one commit it must be -2.

Change-Id: I7838d70652b81e3451f4cda46e309ca6147eb3ca