Warn that project admins are not removed during daily synchronization
This is part of story #9396 have project members automatically synchronized with LDAP

During the daily synchronization or when we bound manually project
members to an LDAP group, the users are now removed if necessary from
project members.

Project administrators are kept and an entry is added in the project history
to say that this user is not removed.

Change-Id: I7cf818204dbdcccd0a4de520d348ab35c625adc1