Performance improvement for the kanban cumulative flow REST route
This is part of story #9463 see cumulative flow diagram

The artifact permissions cache is now common between all the
Tracker_Artifact objects.
On a local dev instance using a kanban based on a tracker of 517
artifacts and a user that is neither a tracker admin or a site admin, I
got a ~75% performance improvement when requesting 14 points. The
performance improvement is due to the reduction of SQL queries executed.
The memory usage is equivalent before and after the optimization.

Change-Id: I9f8c87fcc912997a078c63c99add54d44162033d