Fix request #9162: Escape JSON unparseable characters in REST routes
By default, Restler encodes unicode literally (try to convert \u2260 to
≠). However it breaks the response. For example if someone writes in the
title/description of an artifact "Fix issue with \u2260", it is
converted to "Fix issue with \≠" which 1) is not conform to what the
user wrote and 2) breaks clients.

This options invalids the default behavior and the api explorer is happy
again.

Change-Id: Ia78e0f21cd7d8664d5e52daef1f5814f812f925a