Add ability to change the folder of an artifact
story #9514: change an artifact from one folder to another

On a project that contains tracker folders, take an artifact and in the
artifact link section, you can change its folder.

The _in_folder link has precedence over other types of links:
* If user tries to choose a folder and the folder is already linked with
  another nature (_is_child for example), then the nature will change to
  _in_folder (the link disappear from the table).
* If user tries to add a link (_is_child for example) whereas there is
  already a link _in_folder, then the _is_child will be removed.

The folders are not displayed as a hierarchy yet.

Change-Id: I160d1406a74b7f1cae57cd2be5d791f682b69e6e
TODO: update the value in db