Link to folder is not copied
story #9514: change an artifact from one folder to another

Go to an artifact that belongs to a folder. Copy this artifact. You
should see a warning saying that the link won't be copied.

Go to an artifact that does not belong to a folder. Copy this artifact.
You should NOT see a warning saying that the link won't be copied.

Change-Id: I7f246377ec4381fe6745348e103936b89460f407