request #9466: Enforce that plugins are updated in the same time than Tuleap core
Packages can not be updated if they have not been updated at the same moment than Tuleap core.

This fix the regression introduced by commit 64da72491be75970c1d125eb0f4fe2eeee930f06.

We need to set a release number for the plugins each time Tuleap (core or plugin) is
updated to be sure that the plugins are in sync with the core and to force RPM to see
a change in the plugin packages when Tuleap is updated. If we do not modify the release
number RPM will not see that the requirements of a plugin has changed and it will not be
able to resolve the dependency (you want to update tuleap but tuleap-plugin-xxxxxx requires
an older version of tuleap so we can not update).

The use of _ in the release number seems to be what the Fedora package guidelines tends to
https://fedoraproject.org/wiki/Packaging:Versioning#Post-Release_packages

Force plugins/extensions to have a hard dependency to a core package is what is done
to some packages like the php* ones in the Fedora/CentOS environnement. Take for
example the specfile for PHP 7.1 in the repo of Remi Collet:
https://github.com/remicollet/remirepo/blob/master/php/php71/php71.spec#L267

Change-Id: I2e0f743097f0518f3c679e3d233b1d4b317ecb83