Display hierarchy of folder in read only
story #9514: change an artifact from one folder to another

When you are viewing an artifact in read only mode, if this artifact is
in a folder that has a deep hierarchy, you should be able to see this
hierarchy so that you don't loose context.

The hierarchy is displayed the same way as in Folder column in Folder
tab.

Change-Id: Ifd94d980161064d005050e9ecfb0e96f034b8393