Use tlp-tabs component for the doc navigation
Part of story #8118: TLP

This commit split the documentation into smaller pages.
So that it is easier to work (faster to load, less cognitive overload).

Now to add a new section, you just have to create flat example.html,
doc.html with a manifest.json, and edit the manifest of the parent
section.

Change-Id: I21870c47572018d6e4be2a7885f25ee8400da040