Tracker v3 is no more instanciated for new project
This is part of request #9579 Do not use deprecated services in new projects

If the newly created project is able to use the tracker plugin, then the legacy
tracker service will not be instaciated for this project.
If there are some trackers that need to be duplicated, they are
converted in tracker v5.

Change-Id: Ie78ec75acd757accbf6796628a66f985ede81f1b