request #10195: Use a module-loader in kanban
Sorry for the giant review, if you have ideas on how to split it, let me
know.
This adds a first "component" for the cumulative chart d3 code that is
directly imported in kanban. Instead of adding to a global object, code
using it will need to import it and use a module-loader.
Next steps could be to only import parts of d3 or lodash or moment that
are actually needed.

How to test:
- bower prune; bower install (we still need bower for the modal and
angular-ui-bootstrap)
- yarn (or npm) install in kanban dir
- yarn run build in tuleap root

- yarn run build should still:
- build the SCSS files to bin/assets/kanban.css
- build and minify the JS files to bin/assets/kanban.js
- copy all used images to bin/assets/

- yarn run test should run unit tests once
- yarn run coverage should run unit tests and show test coverage html on
your browser
- yarn run watch should:
- watch js files and rebuild, run unit tests and extract gettext
- watch and build SCSS files

Some of the things that were done in gruntfile have not been migrated:
- adding a banner to the generated minified file
- using ng-annotate. I don't think it's necessary as we use $inject
everywhere.

Change-Id: Ib292bad96febdbe3cf0d77bfd556187f96a12f70