request #9591: Show attachment directly in the changeset
When a user add files to an artifact, it would be great to directly download
the files from the changeset and not have to go back in the attachment field
and retrieve the added files.

Furthermore, if attachment is an image users should have a preview.

Deleted attachment should not be downloadable via changesets though.

Change-Id: Ifdba4927730a01e20fb327d049f71b6272063817