style the user details
Part of story #9590: Get rid of Bootstrap in site admin / users

Go to admin » users, and click on a user. You see user's details, user's
access information and user's projects.

If no projects, you see an empty state.
If admin of projects, you see an icon saying that the user is admin of
them.

If platform uses ldap, then the ldap id is displayed in properties.

Depending on sys_user_approval config, the purpose of the user is
displayed (text that is entered during self registration).

When you click on [change password] button you are redirected in another
page (which probably displays some errors). This will be managed in
another changeset (usage of a modal).

Change-Id: I1f9ebec8cae93f2d182546fbbc4dabe73eed09ea