request #9515 Fatal error while listing permissions of a user group
When a user group has specific permissions on a FRS release, this
release is listed in project » admin » user groups » le_group »
permissions.

However, as soon as the release is deleted there is a fatal error
appearing on the permissions page for the user group.

Bonux: a little glossification of the table.

Change-Id: I6a7037cb83be96e9689e2e26c52d96c2df6fe848