Style pending users
Part of story #9590:  Get rid of Bootstrap in site admin / users

Add style when the pending users list is full. This page is also used
for the list of users pending email activation.

You should play with the following settings to see the changes in
approval buttons: sys_user_approval and platform allows restricted.

Change-Id: I13d926be067e5955fe844a26b0bdc7033d268fad