Git > Gitolite style
This is a part of story #9585 Get rid of Bootstrap in site admin / plugins

To test: on Git > Gitolite configuration, type a project name in the input and click on "Regenerate configuration" button. You should have a page refresh and a blue alert, info type.
If you don't type a project name, you'll have a red alert, danger type.

Change-Id: I0fdabbdd17dcadc9f0b0ce206736b6597302800a