Git > Display repositories for a mirror
Part of story #9585: Get rid of Bootstrap in site admin / plugins

Go to Git » Mirrors.

If a mirror does not have any repositories, then an empty property is
displayed "None".

Else the number is displayed (always in plural). It is a link, when the
user click on it then a modal opens and display the list of repositories
(grouped by project).

This introduces in TLP the notion of tlp-modal-subtitle, and it fixes
also the rendering of the modal structure in the documentation.

Change-Id: I038342564b5bd184c29f104e85cfead59261d46b