Style the admin » projects » project information
Part of story #9589: Get rid of Bootstrap in site admin / projects

The button [Manage Access] in the mockup is moved in a dedicated pane
(see latest wireframe in the story). This pane display the access level
of the project with a corresponding icon. Play with global plateform
access (restricted users) and project access to see the result.

There is an additional pane that displays missing links: project admin,
project members, disk usage (if statistics plugin is available).
Therefore there is no more [project members] in the tabs bar.

When admin resend email instructions, then a feedback is displayed (before the
message was displayed in a nearly blank page).

Change history is moved in a dedicated tab, as per the mockup. However this tab
leads to an error, it will be done in another changeset.

All custom fields (see sidebar "custom project fields") are displayed in a pane
section, if any.

Change-Id: Ic3786ca4e972330ec217d583830adb07c0904ff9