Introduce project history
Part of story #9589: Get rid of Bootstrap in site admin / projects

When you go to admin » projects » project » change history, you see the list of
all events that occured on the project. This list is paginated.

When you go to admin » pending projects » Edit project details » change
history, then the change history is empty (since the project is still pending).

Please note that the filter section is incomplete and doesn't work. It will be
done in a dedicated commit.

Change-Id: I31b6638141df5f9655a8a5a98ef93042181dc6e5