BurningParrot cannot be selected anymore
Part of story #9200: Get rid of Bootstrap in site admin

BurningParrot does not need to be selected to be used. And if it is selected
then we still display most of Tuleap using FlamingParrot. Therefore we must
remove it from user preferences page.

Developers who worked on BurningParrot still have BurningParrot in their
preferences in db but it will be overidden the next time they update their
preferences (no need for a forgeupgrade for this).

Change-Id: I021df156ffec2bf4b7aef0f1e0c6e714a6446ff6