SVN authentication cache can be configured
This is part of request #8756 Revoked token should no longer be usable

You can configure the number of credentials kept by repo and how long a cached credential
can be used.

For now, the cache lifetime is configurable but does nothing.

Reviewers, do not forget to redeploy (or symlink) Tuleap.pm to
/usr/share/perl5/vendor_perl/Apache/Tuleap.pm

You can change default values in the database but you need to refresh the SVNRoot file
by hand

Change-Id: I8b0d6922d3654e71cdef8b2936ac24472100929f