Revert "System events are not ready yet for BurningParrot"
Part of story #9588: Get rid of Bootstrap in site admin / utilities

This reverts commit c832f5f26707c81da24e528a866b26237f60073c.

It only re-applies a previous work already done and merged in the past.
It uses tlp-badges, tlp-tables, tlp-pagination, tlp-panes.

It does not respect at all the mockup, the sidebar is missing, it is
just a starting point to not lose time.

Change-Id: I59946ab426a78bdd589a1c3c6e0709bf32cee13b