story #9734 Directly go to an item by searching its reference
If you are in a project that contains a story #11. Anywhere on the project you
can enter "story #11" in the search box (in the navbar) and you will be
redirected directly on the artifact. This works also with "art #11" or
"artifact #11".

If you go to another project, then "story #11" won't work, but generic
references "art #11" and "artifact #11" will work.

This works also for free with doc/document, …

If the search terms are not the exact references "story #11 and cie", then the
regular search will be performed.

It also works with custom reference created with the import
project script.

Change-Id: Ibf2e38a625311fc637e6b0f0a6c74dc7c2eda0d5