Group system events by date
Part of story #9588: Get rid of Bootstrap in site admin / utilities

In order to produce a more readable table for system events, we now
group events by creation date. The creation/start/end time come back.
The parameters column is less prominent and the [replay] button for
events in error is moved in the modal.

TLP framework is modified: usage of sections in tables forces us to use
multiple tbody, so that the row striping is not screwed when there is
only one row in a section.

Change-Id: I5a7be476066bb3ef6b3e7bbe2fa51a1722550a11