request #9669 Field changes not displayed in last changeset comment
In some cases (not yet clearly defined), the diff for the last
changeset is not displayed.

What I found is that the value for the last changeset is previously
poorly set, so the diff cannot be done. To fix this, we now force to
check in the DB the artifact value as done for email notification
(see Tracker_Artifact_Changeset.class at line 1104)

Change-Id: I54ae0c3c085773018e0bb18176a4f21f44973263