request #9689: Link to avatar or added attachment is wrong
This commit fix the following issue:

When I create a new follow-up comment, my avatar is not displayed in the
email notification.
When I add an attachment, the link to the attachment is wrong.

This is because we are using relative URLs

Please note that if you are using a dev environment, then you still
won't have avatars in gmail since google serves as a proxy to provide
inline href/src. And this proxy don't know how to reach
tuleap-web.tuleap-aio-dev.docker. In the developer console you will see

  https://ci6.googleusercontent.com/proxy/GmW…0-d-e1-ft#https://tuleap-web.tuleap-aio-dev.docker/users/nicolas/avatar.png
  Failed to load resource: the server responded with a status of 404 ()

wherease before this patch you were seeing

  https://ci6.googleusercontent.com/proxy/GmW…0-d-e1-ft#https:///users/nicolas/avatar.png
  Failed to load resource: the server responded with a status of 404 ()

Change-Id: Idb631a6be05764378dc4a12f5efb72d2ec593df1