request #9693: Display rounded avatars in email notifications
When an artifact notification is sent:
- the avatars are rounded
- the border radius on follow-up comments is reduced to 4px

In order to test that I modified maillog plugin so that we can test more
easily the rendering of a mail in the browser. It now uses BurningParrot
and we can open the mail whereas before the information were barely
usable.

It is far from perfect (inline style, inline js) however please consider
that this is a plugin for dev and IMHO we don't need to pass much time on
it.

Change-Id: I8241aea0d2536455d51f511387f3f3e305d8c12d