request #9730: Simplify artifact priority for artifact creation
Insertion at the end of the linked list is the same as insertion at the middle.
However we mix linked list and rank management :(

This means that when an artifact is created, and we put it at the very end of
the list, we also try to update all ranks that are below the… end of the list.
This is useless and has a negative impact on performances for big tables (eg:
500'000 artifacts).

The quick solution, proposed here, is to rewrite the way we put artifacts at
the end.

Change-Id: Ib97ff8e05a7da7b0fe4fb33901ac37d0626ffebf