Correct burndown values in past
request #9732 Burndown chart is not reflecting correctly the remaining effort

Given a burndown with past values,
When values are switch from manual to autocomputed (the same day or not)
The burndown should take the correct values in account

To test it correctly:
On a past same day, change an autocomputed value with a manual value,
then switch bach to autocompute mode.
Then generate cache aigain. Burndown values should be correct

Change-Id: I53ce873665cb467aab8d065b706cc5acd6df3216