Import patterns
This is part of story #9534: import fine grained permissions

How to test this:

In a project.xml file with a git repository export, update the
permissions tag as follow:

<permissions>
...
  <fine_grained enabled="1" use_regexp="1">
     <pattern value="*" type="branch" />
     <pattern value="branch[0-9]" type="branch" />
     <pattern value="*" type="tag />
  </fine_grained>
</permissions>

After importing the project in the Settings > Access control,
the Git repository should use fine-grained permissions with regexp,
and the three patterns must be imported. If you disable regexp usage,
the second pattern must not be import.

Ugroups definition for these patterns will be done in a next commit.

Change-Id: I695e639e7028d26f09de9ed91c788fe303e05bde