request #9759: Developpement environnement uses Gerrit 2.12
Gerrit now tries to check the validation of email addresses which prevent
us to use the API to add the email use by the Git in the tuleap-web container.
Since there is no way to disable this behavior we now allow to forge committer
identity when pushing to refs/meta/config as administrator. This is good enough
in a dev environnement and should not be an issue in production environnement
since the domain name should be considered valid.

Change-Id: Ifadac6ffea1cc9d601a75dc04bb4b5c94cc43f48