request #9769: Be able to build Docker images running Tuleap unit tests again
Remove dead code doing something impossible on real Tuleap instances

We need to be able to update our Docker images running unit tests. They are
currently running on a more recent version of Git coming from RPMForge repo and putting
the Git binary into /usr/bin/git.

RPMForge is no more available so I need to remove this dependency but doing that break the unit
tests since we expect /usr/bin/git to be more recent than the version provided on CentOS6.
Anyway, the code expect a behavior that could not happen on a real server.

There is a lot of dead code in this GitDriver class and the remaining code is only called from
the Git_Exec class (or almost), we can probably do a lot of cleanup/refactoring but this is
a battle for another day.

Change-Id: Ic99d6df84f8103a0ef76c679140079788272a78e