Add fields information in errors
This is a part of story #8858 query tracker with expressions

To test: If you enter a field which doesn't exist or aren't
supported (other that Tracker_FormElement_Field_Text) then
a feedback error is displayed with the associate field.
If you enter several fields non existing or non supported then
a feedback error is displayed on the first field.

Change-Id: I7f4ef23f4bb0976907cf3da473f1d2778078fae7