request #9772: Remove the option 'Stay in HTTPS after login' on the login page
This is something that should not be the responsability of the end user. If we can
access the login page in HTTPS there is no reason to explicitely switch back to
HTTP after the login.

The script DocmanUploader.pl is completely removed by this commit. It contains references
to the option stay_in_ssl and is not usable since the moment of a CSRF protection has been
introduced in the login form.

Change-Id: I3009bc077bb089aa53d5ab589d05042c9d404766