Remove unused exception
This is part of story #9534: import fine grained permissions

For now, when a ugroup defined in Git XML content does not exist,
the script is killed by a PHP Fatal error. To avoid this, we do not use
anymore the exception that was never catched and we log the fact that
the ugroup does not exist.

Change-Id: I9378c82b1c7df80e15714dd68306d875ccfef7e7