Handle expert query during export/import XML
This is a part of story #8858: query tracker with expressions

The expert query (and the fact that the report is/isn't in expert mode) is
exported in XML. If you import this XML structure, then the expert query is
imported.

Old tracker XML structures can still be imported, they won't have expert query
defined.

Change-Id: Id0608f398f34713efc6fc24c9eb30c72cd4b7271