request #9784: Clean up Tuleap Query language grammar
This is part of story #8858: query tracker with expressions

Expert query mode does not throw a notice anymore when trying
to use a not yet implemented feature

Not used or too permissive elements has also been removed from the
grammar. This has a nice side effect to simplify the generated
parser (186KB -> 47KB).

Change-Id: Idbc2b7c325c88cff7805185c213e337573546db6