Refactoring: do not pass unneeded parameters to visitors
This is a part of story #8858 query tracker with expressions

No functional changes.

PHP does not let us to have both proper Parameters and Visitor inheritance.
Therefore we remove the contract definition in the Visitor so that we
can vary the parameters given to each visitor. That way we do not need
to pass not needed parameters to everyone.

However this means that every time we create a visitor, we will need to
guess the needed functions to parse the tree, but we think that it is
the best wow as of today given the PHP limitations.

Change-Id: I46475a95fedbaa88a827df186fe55f26a1462ad6