Invalid query means no results
This is a part of story #8858 query tracker with expressions

If the query has a syntax error, I should not get any results.
If the query has an unknown field, I should not get any results.
If the query has a not supported field, I should not get any results.
If the query is empty, I should get unfiltered results (like in normal
mode)
If the query is valid, I should get matching results.

Change-Id: I721f93d0ad9d356850d2408a9688d27752ea0517