Refactoring - Introduce new object for package permission check
part of request #9802: Access to release changenote doesn't work when release status is hidden

This part only introduce new object to centralise and determine
who can access to it. It should not break anything in FRS:
In FRS try to create a hidden and a visible package and access it with two
users (one who can see it, one who can't).

Next commit will do the same for relase and correct the initial bug.

Change-Id: I1505caa8b143ab3c6f5c02cf864dc2ddcb2ad899