request #9852: Repository can not be accessed through the web UI if the parent repository has been deleted
You can reproduce the issue with the following steps:
1. Create a repository A
2. Fork the repository A into a repository B
3. Delete the repository A
4. Access to the repository B

This fix the issue in pullrequest plugin but this contribution
must be completed by a contribution to the git plugin to have a clean
removal.
Even with the contribution to the git plugin, this local fix is still
interesting. The called method explicitly say that null can be returned
so we should check that is not the case before using the returned object.

Change-Id: Ib8c61eb9e955d7c21bde6b8b3f3ffcc4035bd4f9