request #9867: No need to know if a artifact link type is used when we display the submit new artifact page
When we display the Submit new artifact page, the selectbox for
ArtifactLink type is populated with available natures. However when we
are fetching those natures, we are fetching if they are used. This
information is needed for site admin (in order to disallow the deletion
of the nature) but we don't use/need it for the end user.

Change-Id: I49bd1d75940bf1dffe7c4d612b93f6811a0c9f4a