Sort alphabetically field names in TQL facilitators
Part of story #8859: TQL search on date & numeric fields

Both in the allowed fields selectbox and in the autocompletion, the
fields are sorted so that it is easier to retrieve the information.

Furthermore, in order to reduce the flickering effect on the load of the
page, the textarea has the same height as the codemirror editor.

Change-Id: I8a384a7bf5ff0e2837317d5ba84d3fdee8ed2989