Refacto: simplify date validation
This is a part of story #8859 TQL search on date & numeric fields

By considering that we always get a collection of values to validate for a date
field, we can simplify the current implementation by having only one object to
extract values and one object to validate values.

date = now() => collection of 1 value: [now()]
date between(now() - 1y, now()) => collection of 2 values: [now() - 1y, now()]

On the bonus side, we don't have anymore a "ghost" method visitBetweenXxx that
was empty on a visitor and not on the other.

Change-Id: I77f20654a21e25103804d0f2c4f6f715861589e7