request #9912: Insert field at the right position in TQL editor
When I select a field in the editor and click on the allowed field name,
I expect to see the selection replaced by my choice.

When I put my cursor at the end, middle or start of the expression, and
click on the allowed field name, I expect to see the field added at the
cursor position.

When the field is inserted, I expect to see a space prefixed if needed,
and no unecessary space added (at the beginning of a line, after a
space).

Change-Id: Ia9275f6fe111830b37b96233f9b667c7025f8a99