request #9915 SOAP route getArtifacts generates a lot of PHP warnings
We can provide a criteria parameter to the SOAP route getArtifacts.
However, if this parameter is not an array, this will generate a lot of warnings:
PHP Warning: Invalid argument supplied for foreach() in /usr/share/tuleap/plugins/tracker/include/Tracker/Report/Tracker_Report_SOAP.class.php on line 68

Given that we do not want to break existing implementation, we will ne typehint this parameter,
but do a check before parsing it.

Change-Id: I58221259e59d00d62e53b71bedb09810c2478ac8