request #9916: Call to an unexisting file in frs confirm download popup
Go to FRS, try to download a file. A popup is shown which generates
errors in /var/log/httpd/error_log

The fix for this is really simple: do not load at all a css file (after
all, it was not loaded since FlamingParrot). The popup window will be
replaced by a modal with BurningParrot, so no need to invest much time
on the styling of it now.

Change-Id: If97028c3b1722ac05743883704b7aeed0ead8f64