request #9927: Site admin should be able to easily configure templates
Until now it is quite complex to reach a template. You have to know its
name or you must browse the project lists.

Add a new sidebar item, in "Configuration" section, in order to list all
templates on the plateform. This new page allows:
- to see the project details
- to edit its services
- to edit its trackers

If the template does not use the tracker service, then the corresponding
button is disabled. If the tracker plugin is not enabled, then there
isn't any button at all.

The "Configure predefined services" sidebar item is removed, it was
targetting only the default template (100). The "Configure predefined
trackers" sidebar item must be kept as, in tracker creation form, we
still use the default site template to populate tracker templates.

Change-Id: I19dd56a55484a600f0f24f92a31247e841c454de