Restrict a Gerrit server
This is part of story #9634 Restrict Gerrit server by project

We can now restrict and unrestrict a gerrit server.
Projects that have at least one repository migrated on this gerrit server are
automatically added in the restricted projects list.

This restriction is not taken into account for now.

A refactoring with dedicated objects will be done in a next commit.

Change-Id: I7c599d91f698d27c329ac0a7eab737f103b79ffa