request #9944: Limit exposure of the referrer to protect user's privacy
The end goal is to not leak what the user is doing when he is redirected outside
the Tuleap instance or external sources are included inside the page rendered by the
Tuleap instance.

The referrer-policy header is still quite new and the support from the browsers
not that good so we provide fallback policies as a best effort.

Change-Id: Ic3cecf6a6679cca189783cc20774fbbb746709db