Remove broken default values of the RSS and image widgets
The widget image uses as an example an external URL: we have no insurance
the user can access it and it is an HTTP URL which is going to generate
mixed content errors (it's 2017, HTTPS must be the default).
The RSS image uses an URL that does not work at all anymore.

This is a first step to solve the request #9946.

Change-Id: Iaaa364c5ffe6945d652069a95a14009aa96857ea