Do not potentially leak the origin if the destination is less secure
The first contribution has potentially introduced a policy weaker than
the default one (no-referrer-when-downgrade). There is no good reason to
give the origin if we go to a less secure destination if even the default policy
does not do it.

This contribution closes request #9944: Limit exposure of the referrer to protect user's privacy

Change-Id: I72f03dce1e319f0fd97465bc555a0cf82a456024