Tracker global notifications table has FlamingParrot L&F
Part of story #9892: notify group of people in addition to individuals

Go to tracker » admin » notifications

The table that list global notification should have the standard FlamingParrot
look 'n feel (see mockup in the story).

This commit only target the table, the inside of the table will be done later.

Change-Id: Ic71a8c3395bc317495015982a0de7e5a53d55755