Create table to notify ugroups in git
Part of story #9892: notify group of people in addition to individuals

Run forgeupgrade, hack your database and insert an entry in the table
plugin_git_post_receive_notification_ugroup. You need a repository id,
and a ugroup_id (dynamic or static).

Then push a commit in the given repository. The members of the ugroup
should receive an email.

No changes from the user interface.

Change-Id: I78061dbaefbf83902943c4186428ce06350f4e79