User group can be removed from notifications in git
Part of story #9892: notify group of people in addition to individuals

User groups that are added manually in the database can be removed from the
UI. When the user group is removed, an entry is added in the project history.

The user group cannot be added from the UI yet.

This commit introduces deprecated method in order to easily have a
translated group name. Previous implementations based on ForgeUGroup to
have nice looking name seem to be wrong (git permissions, git
notifications, …?) since a ForgeUgroup !== ProjectUgroup – by reading the
implementation, a ForgeUgroup is a user group that does not belong to
a project (group_id IS NULL). There is a dedicated task that will
refactor a little bit and will remove the added methods.

Change-Id: Ia30d299506e28f477ff4f87fdb1420b9288f2019