Deal with ugroup deletion
Part of story #9892: notify group of people in addition to individuals

Create a ugroup in a project
In database, add this ugroup as to be notified for two repositories for
this project (in plugin_git_post_receive_notification_ugroup table).
Assert that on the UI you have the ugroup listening to the two
repositories.
Delete the ugroup.
Assert that the ugroup are not anymore in the notifications table for
the two repositories.

Change-Id: I66ba0deb4c6fa3b36c70330e786ec9ba905fff13