Refactoring: cleaning of mails to be notified in a dedicated object
Part of story #9892: notify group of people in addition to individuals

This is a first step to be able to remove user from notifications (in a
next commit) when they are not anymore member of the private project.

No functional changes. Given the email address of a user is added to a
repository notifications, when the user is removed from the (private)
project, then the email is removed from the notifications table.

Change-Id: I4fca6c592183dc9e8e3a6890345ab21df1e43a64