Add users or ugroups to notifications table in git
Part of story #9892: notify group of people in addition to individuals

You should be able to add emails, users or ugroups in a git repository
notifications settings.

Users and ugroups are autocompleted.

For ugroups, only project members, project admins and ugroups the user is
member of are returned, so that we do not leak sensitive information
(other customers name for example).

The project history is updated accordingly.

Change-Id: If5a53bc27bd0e52a75af7f738c300998b60e7ce9