request #9970: Tuleap can generate the SSH authorization keys file for Gitolite
To test it you have to:
 - use gitolite3
 - deploy the new sudoers file gitolite3-replace-authorized-keys
 - insert the following entry into the new table: INSERT INTO plugin_git_global_parameters(name, value) VALUES ('authorized_keys_managed', 'tuleap');

New instances will use this mode by default, a next contribution will add a way for existing
instances to switch over this new management mode. The switch will however be one-way, once
Tuleap manages the keys it will not be possible to go back the Gitolite management.

Change-Id: I4556378e7f8ba5780f81ad011b27ee730c69a3d1