Remove hooks.mailinglist from gitolite conf
Part of story #9892: notify group of people in addition to individuals

The `config hooks.mailinglist` is not anymore used for git notifications,
it should be removed from the gitolite repository configuration.

Furthermore we remove triggers for repo update when emails are added to or 
removed from the table since 1) it is useless and 2) already done for the 
mail prefix.

No functional changes. Notifications are still send after a git push to
emails, users, ugroups.

Change-Id: Ic4f295b00dcca21aa29adac46165248ef7281de8