request #10015: Mitigate SHA-1 collision in SVN
This is a first mitigation based on the known SHA-1 collision.
A second step is going to be to integrate a checker for the
collision attack published on SHA-1 [1] to have a more general mitigation.

To test this contribution with SVN core, to do not forget to redeploy
codendi_svn_pre_commit.php. You can find 2 files exploiting the known
SHA-1 collision attached to the request.

[1] https://marc-stevens.nl/research/papers/C13-S.pdf

Change-Id: I928e00ddb7080afb8acb926baee47099169746d8